\name{convexhull-package}
\alias{convexhull-package}
\alias{convexhull}
\docType{package}
\title{
  Package to obtain the convex hull of a set of points.
}
\description{
  This package implements the Jarvis March algorithm for obtaining the convex
  hull of a given list of (x,y) co-ordinate pairs.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Robert Lambert: r.lambert1@lancaster.ac.uk

\keyword{convexhull}
