% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convexhull-package.R
\name{convexhulls}
\alias{convexhulls}
\title{\pkg{convexhulls} draws convex hulls}
\description{
\pkg{convexhulls} draws convex hulls
}
\author{
Mark Heckmann
}
\keyword{convexhulls}
\keyword{package}

