% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imports.R
\name{recode3}
\alias{recode3}
\title{recode function from likert package}
\usage{
recode3(x, from, to, to.class = NULL)
}
\arguments{
\item{x}{The vector whose values will be recoded.}

\item{from}{The old values in x to be recoded.}

\item{to}{The new values.}

\item{to.class}{an 'as.' function representing the desired vector type (i.e.
as.character, as.numeric, as.logical, as.numeric).}
}
\description{
Allows vector input of recodes.
}
\keyword{internal}

