\name{convexhull-package}
\alias{convexhull-package}
\alias{convexhull}
\docType{package}
\title{
  Find the convex hull of a set of points
}
\description{
  Finds the convex hull of a set of points using an RCpp implementation of the Jarvis March algorithm.
}
\details{
  See ReadME.md
}
\author{
Graham Burgess

Maintainer: Graham Burgess
}
\references{
  None
}
\keyword{ package }
\seealso{
  
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
